Revert "Resources"
