+++ src/components/ApplicationViews.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import { withRouter } from "react-router";
-import { Route, Redirect } from "react-router-dom";
+import { Route } from "react-router-dom";
 
 import Home from "./Home/Home";
 import Builds from "./KeyboardBuilder/myBuilds";
@@ -9,8 +9,8 @@ import BuildForm from "./KeyboardBuilder/buildForm";
 import EditBuild from "./KeyboardBuilder/editBuilds";
 import Login from "./login/login";
 import Register from "./login/register";
-import SharedBuilds from "./KeyboardBuilder/Shared/sharedBuilds";
-// import SharedDetails from "./KeyboardBuilder/sharedDetails"
+import SharedBuilds from './KeyboardBuilder/sharedBuilds'
+
 class ApplicationViews extends Component {
   state = {
     build: [],
@@ -44,8 +44,6 @@ class ApplicationViews extends Component {
         this.setState({ build: editedBuild });
       });
 
-  isAuthenticated = () => sessionStorage.getItem("User") !== null;
-
   componentDidMount() {
     KeyboardManager.getAllBuilds().then(allBuilds => {
       this.setState({
@@ -69,67 +67,33 @@ class ApplicationViews extends Component {
           exact
           path="/my-builds"
           render={props => {
-            if (this.isAuthenticated()) {
-              return (
-                <Builds
-                  {...props}
-                  build={this.state.build}
-                  deleteBuild={this.deleteBuild}
-                />
-              );
-            } else {
-              return <Redirect to="/login" />;
-            }
+            return (
+              <Builds
+                {...props}
+                build={this.state.build}
+                deleteBuild={this.deleteBuild}
+              />
+            );
           }}
         />
 
         <Route
           path="/my-builds/add-build"
           render={props => {
-            if (this.isAuthenticated()) {
-              return <BuildForm {...props} addBuild={this.addBuild} />;
-            } else {
-              return <Redirect to="/login" />;
-            }
+            return <BuildForm {...props} addBuild={this.addBuild} />;
           }}
         />
 
         <Route
           path="/my-builds/:buildId(\d+)/edit"
           render={props => {
-            if (this.isAuthenticated()) {
-              return  <EditBuild {...props} updateBuild={this.updateBuild} />;
-            } else {
-              return <Redirect to="/login" />;
-            }
+            return <EditBuild {...props} updateBuild={this.updateBuild} />;
           }}
         />
 
-        <Route
-          path="/shared-builds"
-          render={props => {
-            return <SharedBuilds {...props} build={this.state.build} />;
-          }}
-        />
-
-        {/* <Route 
-        path="/shared-builds/:buildId(\d+)"
-        render={(props) => {
-          let build = this.state.build.find(build => build.id === parseInt(props.match.params.buildId)
-          )
-          if(!build) {
-            build = {id: 404, name: "404", location: "Page not found"}
-          }
-
-          return(
-            <SharedDetails 
-            build={build}
-            {...props}
-            builds={this.state.build}
-            />
-          )
-        }}
-            /> */}
+          <Route path="/shared-builds" render={props => {
+            return <SharedBuilds {...props} build={this.state.build} />
+          }}/>
 
         <Route
           path="/register"
