+++ src/index.css
@@ -5,8 +5,6 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  background-image: url(./components/images/office.jpg);
-  background-size: 100%
 }
 
 code {
